Simplify gestaltd core package structure without behavior changes.
Split oversized agent and IndexedDB test stub files, deduplicate provider
capability detection, and consolidate Actor, subject context, and catalog
annotation types behind aliases so downstream imports stay stable.
