--- gestaltd/core/agent/agent.go
@@ -1,675 +1,0 @@
-package agent
-
-import (
-	"context"
-	"fmt"
-	"strings"
-	"time"
-
-	"github.com/valon-technologies/gestalt/server/core"
-)
-
-type ExecutionStatus string
-
-const (
-	ExecutionStatusPending         ExecutionStatus = "pending"
-	ExecutionStatusRunning         ExecutionStatus = "running"
-	ExecutionStatusSucceeded       ExecutionStatus = "succeeded"
-	ExecutionStatusFailed          ExecutionStatus = "failed"
-	ExecutionStatusCanceled        ExecutionStatus = "canceled"
-	ExecutionStatusWaitingForInput ExecutionStatus = "waiting_for_input"
-)
-
-func ExecutionStatusIsLive(status ExecutionStatus) bool {
-	switch status {
-	case ExecutionStatusPending, ExecutionStatusRunning, ExecutionStatusWaitingForInput:
-		return true
-	default:
-		return false
-	}
-}
-
-type SessionState string
-
-const (
-	SessionStateActive   SessionState = "active"
-	SessionStateArchived SessionState = "archived"
-)
-
-type Actor struct {
-	SubjectID   string
-	SubjectKind string
-	DisplayName string
-	AuthSource  string
-}
-
-type SubjectContext struct {
-	SubjectID           string
-	SubjectKind         string
-	CredentialSubjectID string
-	DisplayName         string
-	AuthSource          string
-}
-
-type Message struct {
-	Role     string
-	Text     string
-	Parts    []MessagePart
-	Metadata map[string]any
-}
-
-type MessagePartType string
-
-const (
-	MessagePartTypeText       MessagePartType = "text"
-	MessagePartTypeJSON       MessagePartType = "json"
-	MessagePartTypeToolCall   MessagePartType = "tool_call"
-	MessagePartTypeToolResult MessagePartType = "tool_result"
-	MessagePartTypeImageRef   MessagePartType = "image_ref"
-)
-
-type ToolCallPart struct {
-	ID        string
-	ToolID    string
-	Arguments map[string]any
-}
-
-type ToolResultPart struct {
-	ToolCallID string
-	Status     int
-	Content    string
-	Output     map[string]any
-}
-
-type ImageRefPart struct {
-	URI      string
-	MIMEType string
-}
-
-const SystemToolWorkflow = "workflow"
-
-type MessagePart struct {
-	Type       MessagePartType
-	Text       string
-	JSON       map[string]any
-	ToolCall   *ToolCallPart
-	ToolResult *ToolResultPart
-	ImageRef   *ImageRefPart
-}
-
-type ToolTarget struct {
-	System                string `json:",omitempty"`
-	App                   string
-	Operation             string
-	Connection            string
-	Instance              string
-	CredentialMode        core.ConnectionMode
-	Unavailable           *UnavailableToolTarget `json:",omitempty"`
-	RunAs                 *core.RunAsSubject
-	RunAsExternalIdentity *core.ExternalIdentityRef
-}
-
-type UnavailableToolTarget struct {
-	Reason  string
-	Message string
-}
-
-const (
-	ToolUnavailableReasonReconnectRequired = "reconnect_required"
-	ToolUnavailableReasonNotAuthenticated  = "not_authenticated"
-	ToolUnavailableReasonNoCredential      = "no_credential"
-	ToolUnavailableReasonScopeDenied       = "scope_denied"
-	ToolUnavailableReasonInstanceRequired  = "instance_required"
-)
-
-type Tool struct {
-	ID               string
-	Name             string
-	Description      string
-	ParametersSchema map[string]any
-	Target           ToolTarget
-	Hidden           bool
-}
-
-type ToolRef struct {
-	System                string `json:",omitempty"`
-	App                   string
-	Operation             string
-	Connection            string
-	Instance              string
-	CredentialMode        core.ConnectionMode
-	RunAs                 *core.RunAsSubject
-	RunAsExternalIdentity *core.ExternalIdentityRef
-	Title                 string
-	Description           string
-}
-
-type ResolveToolsRequest struct {
-	ToolRefs      []ToolRef
-	ToolSource    ToolSourceMode
-	CallerAppName string
-}
-
-type ToolSourceMode string
-
-const (
-	ToolSourceModeUnspecified ToolSourceMode = ""
-	ToolSourceModeMCPCatalog  ToolSourceMode = "mcp_catalog"
-	ToolSourceModeNone        ToolSourceMode = "none"
-)
-
-func ValidateMCPCatalogToolRefs(refs []ToolRef, fieldName string) error {
-	fieldName = strings.TrimSpace(fieldName)
-	if fieldName == "" {
-		fieldName = "toolRefs"
-	}
-	for i := range refs {
-		ref := refs[i]
-		system := strings.TrimSpace(ref.System)
-		pluginName := strings.TrimSpace(ref.App)
-		operation := strings.TrimSpace(ref.Operation)
-		connection := strings.TrimSpace(ref.Connection)
-		instance := strings.TrimSpace(ref.Instance)
-		if system != "" && system != SystemToolWorkflow {
-			return fmt.Errorf("mcp catalog %s[%d].system %q is not supported", fieldName, i, system)
-		}
-		if system != "" && pluginName != "" {
-			return fmt.Errorf("mcp catalog %s[%d] must set exactly one of app or system", fieldName, i)
-		}
-		if system != "" {
-			if operation == "" {
-				return fmt.Errorf("mcp catalog %s[%d].operation is required for system refs", fieldName, i)
-			}
-			if operation == "*" {
-				return fmt.Errorf("mcp catalog %s[%d].operation wildcard is not supported", fieldName, i)
-			}
-			if connection != "" || instance != "" || ref.CredentialMode != "" || ref.RunAs != nil || ref.RunAsExternalIdentity != nil || strings.TrimSpace(ref.Title) != "" || strings.TrimSpace(ref.Description) != "" {
-				return fmt.Errorf("mcp catalog %s[%d] system refs cannot include connection, instance, credential mode, runAs, runAs external identity, title, or description", fieldName, i)
-			}
-			continue
-		}
-		if pluginName == "" {
-			return fmt.Errorf("mcp catalog %s[%d].app is required", fieldName, i)
-		}
-		if operation == "*" || connection == "*" || instance == "*" {
-			return fmt.Errorf("mcp catalog %s[%d] wildcard fields are not supported", fieldName, i)
-		}
-		if pluginName == "*" && (operation != "" || connection != "" || instance != "" || ref.CredentialMode != "" || ref.RunAs != nil || ref.RunAsExternalIdentity != nil || strings.TrimSpace(ref.Title) != "" || strings.TrimSpace(ref.Description) != "") {
-			return fmt.Errorf("mcp catalog %s[%d] global ref cannot include operation, connection, instance, credential mode, runAs, runAs external identity, title, or description", fieldName, i)
-		}
-		if ref.RunAs != nil && operation == "" {
-			return fmt.Errorf("mcp catalog %s[%d].runAs requires an exact operation", fieldName, i)
-		}
-		if ref.RunAsExternalIdentity != nil && ref.RunAs == nil {
-			return fmt.Errorf("mcp catalog %s[%d].runAs.externalIdentity requires runAs.subject", fieldName, i)
-		}
-		if ref.RunAsExternalIdentity != nil && operation == "" {
-			return fmt.Errorf("mcp catalog %s[%d].runAs.externalIdentity requires an exact operation", fieldName, i)
-		}
-	}
-	return nil
-}
-
-type Session struct {
-	ID           string
-	ProviderName string
-	Model        string
-	ClientRef    string
-	State        SessionState
-	Metadata     map[string]any
-	CreatedBy    Actor
-	CreatedAt    *time.Time
-	UpdatedAt    *time.Time
-	LastTurnAt   *time.Time
-}
-
-type CreateSessionRequest struct {
-	SessionID         string
-	IdempotencyKey    string
-	Model             string
-	ClientRef         string
-	Metadata          map[string]any
-	CreatedBy         Actor
-	Subject           SubjectContext
-	SessionStart      *SessionStartConfig
-	Workspace         *Workspace
-	PreparedWorkspace *PreparedWorkspace
-}
-
-type Workspace struct {
-	Checkouts []WorkspaceGitCheckout
-	CWD       string
-}
-
-type WorkspaceGitCheckout struct {
-	URL  string
-	Ref  string
-	Path string
-}
-
-type PreparedWorkspace struct {
-	Root string
-	CWD  string
-}
-
-type SessionStartConfig struct {
-	Hooks []SessionStartHook
-}
-
-type SessionStartHook struct {
-	ID      string
-	Type    string
-	Command []string
-	CWD     string
-	Timeout string
-	Env     map[string]string
-	Output  SessionStartHookOutput
-}
-
-type SessionStartHookOutput struct {
-	AdditionalContext bool
-	Metadata          bool
-}
-
-type GetSessionRequest struct {
-	SessionID string
-	Subject   SubjectContext
-}
-
-type ListSessionsRequest struct {
-	Subject    SubjectContext
-	SessionIDs []string
-	State      SessionState
-	// Limit asks providers to apply the cap before returning.
-	// Providers must order sessions by recency before limiting: last turn time, then
-	// updated time, then created time, newest first.
-	Limit int
-	// SummaryOnly asks providers to omit heavy fields.
-	// Exact SessionIDs may still be served by direct lookup rather than projections.
-	SummaryOnly bool
-}
-
-type UpdateSessionRequest struct {
-	SessionID string
-	ClientRef string
-	State     SessionState
-	Metadata  map[string]any
-	Subject   SubjectContext
-}
-
-type Turn struct {
-	ID               string
-	SessionID        string
-	ProviderName     string
-	Model            string
-	Status           ExecutionStatus
-	Messages         []Message
-	OutputText       string
-	StructuredOutput map[string]any
-	StatusMessage    string
-	CreatedBy        Actor
-	CreatedAt        *time.Time
-	StartedAt        *time.Time
-	CompletedAt      *time.Time
-	ExecutionRef     string
-}
-
-type CreateTurnRequest struct {
-	TurnID            string
-	SessionID         string
-	IdempotencyKey    string
-	Model             string
-	Messages          []Message
-	ToolRefs          []ToolRef
-	ToolSource        ToolSourceMode
-	Tools             []Tool
-	ResponseSchema    map[string]any
-	ResponseSchemaSet bool
-	Metadata          map[string]any
-	ModelOptions      map[string]any
-	TimeoutSeconds    int
-	CreatedBy         Actor
-	ExecutionRef      string
-	Subject           SubjectContext
-	RunGrant          string
-}
-
-type GetTurnRequest struct {
-	TurnID  string
-	Subject SubjectContext
-}
-
-type ListTurnsRequest struct {
-	SessionID string
-	Subject   SubjectContext
-	TurnIDs   []string
-	Status    ExecutionStatus
-	// Limit asks providers to apply the cap before returning.
-	// Providers must order turns by creation time, newest first, before limiting.
-	Limit int
-	// SummaryOnly asks providers to omit heavy turn fields.
-	// Exact TurnIDs may still be served by direct lookup rather than projections.
-	SummaryOnly bool
-}
-
-type CancelTurnRequest struct {
-	TurnID  string
-	Reason  string
-	Subject SubjectContext
-}
-
-type TurnEvent struct {
-	ID         string
-	TurnID     string
-	Seq        int64
-	Type       string
-	Source     string
-	Visibility string
-	Data       map[string]any
-	CreatedAt  *time.Time
-	Display    *TurnDisplay
-}
-
-type TurnDisplay struct {
-	Kind      string
-	Phase     string
-	Text      string
-	Label     string
-	Ref       string
-	ParentRef string
-	Input     any
-	Output    any
-	Error     any
-	Action    string
-	Format    string
-	Language  string
-}
-
-type ListTurnEventsRequest struct {
-	TurnID   string
-	AfterSeq int64
-	Limit    int
-	Subject  SubjectContext
-}
-
-type GetCapabilitiesRequest struct{}
-
-type ProviderCapabilities struct {
-	StreamingText             bool
-	ToolCalls                 bool
-	ParallelToolCalls         bool
-	StructuredOutput          bool
-	Interactions              bool
-	ResumableTurns            bool
-	ReasoningSummaries        bool
-	SupportsSessionStart      bool
-	SupportsPreparedWorkspace bool
-	// BoundedListHydration means provider list APIs can apply Limit and SummaryOnly
-	// without hydrating every source record, while preserving the list ordering
-	// contract before applying Limit.
-	BoundedListHydration bool
-	SupportedToolSources []ToolSourceMode
-}
-
-type GetInteractionRequest struct {
-	InteractionID string
-	Subject       SubjectContext
-}
-
-type ListInteractionsRequest struct {
-	TurnID  string
-	Subject SubjectContext
-}
-
-type ResolveInteractionRequest struct {
-	InteractionID string
-	Resolution    map[string]any
-	Subject       SubjectContext
-}
-
-type ExecuteToolRequest struct {
-	ProviderName   string
-	SessionID      string
-	TurnID         string
-	ToolCallID     string
-	ToolID         string
-	Arguments      map[string]any
-	RunGrant       string
-	IdempotencyKey string
-}
-
-type ExecuteToolResponse struct {
-	Status int
-	Body   string
-}
-
-type ListedTool struct {
-	ToolID           string
-	MCPName          string
-	Title            string
-	Description      string
-	Tags             []string
-	SearchText       string
-	InputSchemaJSON  string
-	OutputSchemaJSON string
-	Annotations      core.CapabilityAnnotations
-	Ref              ToolRef
-	Target           ToolTarget
-	Hidden           bool
-}
-
-type ListToolsRequest struct {
-	ProviderName string
-	SessionID    string
-	TurnID       string
-	PageSize     int
-	PageToken    string
-	Query        string
-	ToolRefs     []ToolRef
-	ToolSource   ToolSourceMode
-	RunGrant     string
-}
-
-type ListToolsResponse struct {
-	Tools         []ListedTool
-	NextPageToken string
-}
-
-type ResolveConnectionRequest struct {
-	ProviderName string
-	SessionID    string
-	TurnID       string
-	Connection   string
-	Instance     string
-	RunGrant     string
-}
-
-type ResolvedConnection struct {
-	ConnectionID string
-	Connection   string
-	Instance     string
-	Mode         core.ConnectionMode
-	Headers      map[string]string
-	Params       map[string]string
-	ExpiresAt    *time.Time
-}
-
-type InteractionType string
-
-const (
-	InteractionTypeApproval      InteractionType = "approval"
-	InteractionTypeClarification InteractionType = "clarification"
-	InteractionTypeInput         InteractionType = "input"
-)
-
-type InteractionState string
-
-const (
-	InteractionStatePending  InteractionState = "pending"
-	InteractionStateResolved InteractionState = "resolved"
-	InteractionStateCanceled InteractionState = "canceled"
-)
-
-type Interaction struct {
-	ID         string
-	TurnID     string
-	SessionID  string
-	Type       InteractionType
-	State      InteractionState
-	Title      string
-	Prompt     string
-	Request    map[string]any
-	Resolution map[string]any
-	CreatedAt  *time.Time
-	ResolvedAt *time.Time
-}
-
-type ManagerCreateSessionRequest struct {
-	IdempotencyKey string
-	ProviderName   string
-	Model          string
-	ClientRef      string
-	Metadata       map[string]any
-	Workspace      *Workspace
-}
-
-type ManagerUpdateSessionRequest struct {
-	SessionID string
-	ClientRef string
-	State     SessionState
-	Metadata  map[string]any
-}
-
-type ManagerCreateTurnRequest struct {
-	CallerAppName     string
-	IdempotencyKey    string
-	Model             string
-	SessionID         string
-	Messages          []Message
-	ToolRefs          []ToolRef
-	ToolRefsSet       bool
-	ToolSource        ToolSourceMode
-	ResponseSchema    map[string]any
-	ResponseSchemaSet bool
-	Metadata          map[string]any
-	ModelOptions      map[string]any
-	TimeoutSeconds    int
-}
-
-type ManagerListSessionsRequest struct {
-	ProviderName string
-	State        SessionState
-	// Limit caps the globally sorted manager response. Bounded providers receive
-	// this cap for per-provider listing.
-	Limit       int
-	SummaryOnly bool
-}
-
-type ManagerListTurnsRequest struct {
-	SessionID   string
-	Status      ExecutionStatus
-	Limit       int
-	SummaryOnly bool
-}
-
-type ManagerCancelTurnRequest struct {
-	TurnID string
-	Reason string
-}
-
-// Provider is the authoritative agent data boundary. Read methods for
-// sessions, turns, turn events, and interactions must be served from
-// provider-owned control-plane state and must not require a live execution
-// sandbox, pod IP, cached tunnel, or other transport attachment. Execution
-// methods may require runtime transport and should return an unavailable error
-// when the provider has durable state but cannot attach to the execution
-// runtime.
-type Provider interface {
-	CreateSession(ctx context.Context, req CreateSessionRequest) (*Session, error)
-	GetSession(ctx context.Context, req GetSessionRequest) (*Session, error)
-	ListSessions(ctx context.Context, req ListSessionsRequest) ([]*Session, error)
-	UpdateSession(ctx context.Context, req UpdateSessionRequest) (*Session, error)
-	CreateTurn(ctx context.Context, req CreateTurnRequest) (*Turn, error)
-	GetTurn(ctx context.Context, req GetTurnRequest) (*Turn, error)
-	ListTurns(ctx context.Context, req ListTurnsRequest) ([]*Turn, error)
-	CancelTurn(ctx context.Context, req CancelTurnRequest) (*Turn, error)
-	ListTurnEvents(ctx context.Context, req ListTurnEventsRequest) ([]*TurnEvent, error)
-	GetInteraction(ctx context.Context, req GetInteractionRequest) (*Interaction, error)
-	ListInteractions(ctx context.Context, req ListInteractionsRequest) ([]*Interaction, error)
-	ResolveInteraction(ctx context.Context, req ResolveInteractionRequest) (*Interaction, error)
-	GetCapabilities(ctx context.Context, req GetCapabilitiesRequest) (*ProviderCapabilities, error)
-	Ping(ctx context.Context) error
-	Close() error
-}
-
-type WorkspaceProvider interface {
-	Provider
-	SupportsWorkspaceRequests() bool
-}
-
-type Host interface {
-	ListTools(ctx context.Context, req ListToolsRequest) (*ListToolsResponse, error)
-	ExecuteTool(ctx context.Context, req ExecuteToolRequest) (*ExecuteToolResponse, error)
-}
-
-type UnimplementedProvider struct{}
-
-func (UnimplementedProvider) CreateSession(context.Context, CreateSessionRequest) (*Session, error) {
-	return nil, fmt.Errorf("agent provider create session is not implemented")
-}
-
-func (UnimplementedProvider) GetSession(context.Context, GetSessionRequest) (*Session, error) {
-	return nil, fmt.Errorf("agent provider get session is not implemented")
-}
-
-func (UnimplementedProvider) ListSessions(context.Context, ListSessionsRequest) ([]*Session, error) {
-	return nil, fmt.Errorf("agent provider list sessions is not implemented")
-}
-
-func (UnimplementedProvider) UpdateSession(context.Context, UpdateSessionRequest) (*Session, error) {
-	return nil, fmt.Errorf("agent provider update session is not implemented")
-}
-
-func (UnimplementedProvider) CreateTurn(context.Context, CreateTurnRequest) (*Turn, error) {
-	return nil, fmt.Errorf("agent provider create turn is not implemented")
-}
-
-func (UnimplementedProvider) GetTurn(context.Context, GetTurnRequest) (*Turn, error) {
-	return nil, fmt.Errorf("agent provider get turn is not implemented")
-}
-
-func (UnimplementedProvider) ListTurns(context.Context, ListTurnsRequest) ([]*Turn, error) {
-	return nil, fmt.Errorf("agent provider list turns is not implemented")
-}
-
-func (UnimplementedProvider) CancelTurn(context.Context, CancelTurnRequest) (*Turn, error) {
-	return nil, fmt.Errorf("agent provider cancel turn is not implemented")
-}
-
-func (UnimplementedProvider) ListTurnEvents(context.Context, ListTurnEventsRequest) ([]*TurnEvent, error) {
-	return nil, fmt.Errorf("agent provider list turn events is not implemented")
-}
-
-func (UnimplementedProvider) GetInteraction(context.Context, GetInteractionRequest) (*Interaction, error) {
-	return nil, fmt.Errorf("agent provider get interaction is not implemented")
-}
-
-func (UnimplementedProvider) ListInteractions(context.Context, ListInteractionsRequest) ([]*Interaction, error) {
-	return nil, fmt.Errorf("agent provider list interactions is not implemented")
-}
-
-func (UnimplementedProvider) ResolveInteraction(context.Context, ResolveInteractionRequest) (*Interaction, error) {
-	return nil, fmt.Errorf("agent provider resolve interaction is not implemented")
-}
-
-func (UnimplementedProvider) GetCapabilities(context.Context, GetCapabilitiesRequest) (*ProviderCapabilities, error) {
-	return nil, fmt.Errorf("agent provider get capabilities is not implemented")
-}
-
-func (UnimplementedProvider) Ping(context.Context) error {
-	return fmt.Errorf("agent provider ping is not implemented")
-}
-
-func (UnimplementedProvider) Close() error {
-	return nil
-}
