+++ gestaltd/core/subjects.go
@@ -2,6 +2,31 @@ package core
 
 import "strings"
 
+// Actor identifies who performed an action without credential-scoping fields.
+type Actor struct {
+	SubjectID   string
+	SubjectKind string
+	DisplayName string
+	AuthSource  string
+}
+
+func (a Actor) IsZero() bool {
+	return a == Actor{}
+}
+
+func ActorFromRunAs(subject *RunAsSubject) Actor {
+	subject = NormalizeRunAsSubject(subject)
+	if subject == nil {
+		return Actor{}
+	}
+	return Actor{
+		SubjectID:   subject.SubjectID,
+		SubjectKind: subject.SubjectKind,
+		DisplayName: subject.DisplayName,
+		AuthSource:  subject.AuthSource,
+	}
+}
+
 // TODO(#1823): Add first-class run-as subject and external-identity grant
 // provisioning instead of relying on opaque subject IDs plus separate tuple
 // seeding.
