+++ gestaltd/internal/bootstrap/agent_runtime_test.go
@@ -2840,7 +2840,7 @@ func TestAgentRuntimeListsMCPCatalogToolsForGrantedTurn(t *testing.T) {
 				Title:       "Sync roadmap",
 				Description: "Sync roadmap tasks",
 				InputSchema: json.RawMessage(`{"type":"object","properties":{"taskId":{"type":"string"}}}`),
-				Annotations: catalog.OperationAnnotations{
+				Annotations: core.CapabilityAnnotations{
 					ReadOnlyHint: &readOnly,
 				},
 			},
