+++ gestaltd/services/agents/convert_agent.go
@@ -68,13 +68,7 @@ func subjectToProto(subject coreagent.SubjectContext) *proto.SubjectContext {
 	if subject == (coreagent.SubjectContext{}) {
 		return nil
 	}
-	return &proto.SubjectContext{
-		Id:                  subject.SubjectID,
-		Kind:                subject.SubjectKind,
-		CredentialSubjectId: subject.CredentialSubjectID,
-		DisplayName:         subject.DisplayName,
-		AuthSource:          subject.AuthSource,
-	}
+	return agentwire.RunAsSubjectToProto(&subject)
 }
 
 func agentWorkspaceFromProto(workspace *proto.AgentWorkspace) *coreagent.Workspace {
