+++ gestaltd/services/apps/declarative/schema.go
@@ -7,6 +7,7 @@ import (
 	"slices"
 	"strings"
 
+	"github.com/valon-technologies/gestalt/server/core"
 	"github.com/valon-technologies/gestalt/server/core/catalog"
 )
 
@@ -90,8 +91,8 @@ type topLevelInputSchema struct {
 }
 
 // AnnotationsFromMethod derives MCP operation annotations from an HTTP method.
-func AnnotationsFromMethod(method string) catalog.OperationAnnotations {
-	a := catalog.OperationAnnotations{
+func AnnotationsFromMethod(method string) core.CapabilityAnnotations {
+	a := core.CapabilityAnnotations{
 		OpenWorldHint: boolPtr(true),
 	}
 	switch strings.ToUpper(method) {
@@ -118,7 +119,7 @@ func CompileSchemas(c *catalog.Catalog) {
 		if len(op.Parameters) == 0 {
 			op.Parameters = parametersFromInputSchema(op.InputSchema)
 		}
-		if op.Annotations == (catalog.OperationAnnotations{}) {
+		if op.Annotations == (core.CapabilityAnnotations{}) {
 			op.Annotations = AnnotationsFromMethod(op.Method)
 		}
 	}
